+++ HISTORY.md
@@ -3,6 +3,7 @@ unreleased
 
   * Accept `req` as argument to `parse`
   * Accept `res` as argument to `parse`
+  * Parse media type with extra LWS between type and first parameter
 
 0.0.0 / 2014-06-13
 ==================